Make tests run
