+++ package.json
@@ -66,6 +66,7 @@
   "dependencies": {
     "autoprefixer": "^6.4.0",
     "bootstrap-sass": "^3.3.7",
+    "bootstrap-sass-loader": "^1.0.10",
     "core-js": "^2.0.0",
     "hel-bootstrap-3": "^1.0.0-beta.0",
     "jquery": "^3.1.0",
@@ -74,11 +75,12 @@
     "node-sass": "^3.8.0",
     "normalize.css": "^4.0.0",
     "oidc-client": "^1.3.0",
-    "react": "15.3.0",
+    "prop-types": "^15.5.10",
+    "react": "^15.3.0",
     "react-bootstrap": "^0.31.0",
-    "react-dom": "15.3.0",
+    "react-dom": "^15.3.0",
     "react-redux": "^4.4.5",
-    "react-router": "^2.6.1",
+    "react-router": "^3.0.0",
     "react-router-redux": "^4.0.8",
     "redux": "^3.5.2",
     "redux-actions": "^2.0.3",
