+++ src/components/CallbackPage.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { CallbackComponent } from 'redux-oidc';
 import userManager from '../util/user-manager';
 import _ from 'lodash';
@@ -21,7 +22,7 @@ class CallbackPage extends React.Component {
 }
 
 CallbackPage.contextTypes = {
-  router: React.PropTypes.object
+  router: PropTypes.object
 };
 
 export default CallbackPage;
