+++ test/helpers/shallowRenderHelper.js
@@ -6,7 +6,7 @@
  * @author somonsmith
  */
 import React from 'react';
-import TestUtils from 'react-addons-test-utils';
+import TestUtils from 'react-dom/test-utils';
 
 /**
  * Get the shallow rendered component
